containerd: add initial helper scripts for configuring containerd
configure-containerd.sh will install containerd (1.2.2), and configure it to
support three runtimeClasses: runc, kata-fc and kata-qemu.

It assumes the paths for the kata binaries are at /opt/kata/bin/

Kubelet service file still needs to be updated to use containerd, but
since this is distro specific (or, at least different for ClearLinux),
I'm leaving it up to the user to make this update.  It should simply
update the CRI socket path. If crio was installed, it'd be
s/crio/containerd

Signed-off-by: Eric Ernst <eric.ernst@intel.com>