metrics: report: scaling: symbolise node_util data extraction
The node_util data extraction was somewhat hard wired with numeric
indexing into the JSON imported array. This fairly hard tied the code
to the layout of the JSON. Adding more elements to the JSON would break
the parsing.

Move the data parse out to use the symbolic names (with some coding
to iterate the columns of the data frame, not the rows), to enable
future JSON structural changes, allowing the parsing to still work
(as long as we don't change any element names).

Signed-off-by: Graham Whaley <graham.whaley@intel.com>