Fix no_proxy issue and remove hardcoded box ver
Fixed an issue where no_proxy was not getting set to include the
CIDR being used by the CNI. This resulted in kubeadm join command
trying to go out through the proxy server.
Also removed hardcoded vagrant box version. This change will now
allow the override of the version, or if no override is passed in
it will take the version on the system, or download the latest version.

Closes issue #136