metrics: scaling: Fix R error with nodes>1 scaling data
the tidy_scaling.R code contains a mistake, where it tries to match the
node names into the local_bootdata array. This does not work if we have
more than one worker node and STEP>1, and results in an
'array mismatch' error.

Fundamentally we cannot try to match a single node to a single launch in
that scenario (as we will have >1 pods and probably >1 nodes the pods
launched on), but that information is stored already in the
local_nodedata table anyway.

Fix this by not trying to assign node names in the bootdata table, and
use the local_nodedata node info to generate the graphs.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>