metrics: scaling: only measure inodes for root volume
The stats pod code was collecting the inode count by looking for any
entry in the `df` that started with 'overlay'. This does not work on all
systems (such as under multi-node 'kind'), where there are multiple
overlay filesystems mounted. This results in a multi data bash array
being returned, which then fails the 'math' code later on.

Fix by explicitly only looking for the root partition.

Note, this is still not ideal, as this code will only work for overlay
rootfs mounts, which may not necessarily be the case for all runtimes?
(such as kata). But, this fix is an improvement and works with the
current use cases.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>