metrics: collected: use linear regression for cpu calc
We were using a simple last-first calculation to work out the per-pod
cpu overhead - but, the cpu data is pretty noisy, so it was very easy to
hit a peak or trough at one or either end, and throw the result out.

Try to use a linear regression to find the trend of the slope, and work
out the intercept points for the first and last sample, and thus the
linear cpu reduction.

Update the markdown to place the CPU data notes next to the CPU graphs.

Note, this only really works for workloads that are linear, but, that
was true for the previous method as well, so is no worse. If the
regression model fails, then `NA` will appear in the stats table, which
is probably better than printing untrue data.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>