rook: allow multiple mons per node
On a single node setup, rook will fail to get stuck waiting on OSDs as
monitors by default have an anti-affinity rule that prevents multiple
monitors to land on a single node. This will also be an issue if you are
using this setup of < 4 nodes (1 master and 3 workers) as the master has
the NoSchedule taint. This patch allows multiple monitors on a single
node. When multiple nodes are present, the monitor pods are
automatically distributed.

To allow single node setup to work, the default replicapool replicas has
been changed from 3 to 1. If this setup is used in a production cluster,
we recommend the user create a ceph pool that suits their needs against
using the default pool.

Downside: When the setup is started with a single node and then more
nodes are added, the monitor pods *DO NOT* automatically re-balance.

Fixes: #306

Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>